+++ src/Classes/comps/Component.ts
@@ -1,5 +1,5 @@
-import { Vector2 } from "../utils/types";
-import Renderer from "./Renderer";
+import { Vector2 } from "../../utils/types";
+import Renderer from "../Renderer";
 
 class Component {
   private element: HTMLElement;
